Move all the descriptor parsing/mangling in schema-impl
