+++ crates/meta/Cargo.toml
@@ -38,7 +38,7 @@ prost-reflect = { workspace = true }
 restate-fs-util = { workspace = true }
 restate-meta-rest-model = { workspace = true, features = ["schema"] }
 restate-pb = { workspace = true }
-restate-schema-api = { workspace = true, features = ["service", "endpoint", "discovery", "serde", "serde_schema"] }
+restate-schema-api = { workspace = true, features = ["service", "endpoint", "serde", "serde_schema"] }
 restate-schema-impl = { workspace = true }
 restate-service-client = { workspace = true }
 restate-service-protocol = { workspace = true, features = ["discovery"] }
