+++ crates/schema-impl/tests/pb/change_instance_type/v2/greeter.proto
@@ -1,0 +1,20 @@
+syntax = "proto3";
+
+import "google/protobuf/empty.proto";
+import "dev/restate/ext.proto";
+
+package greeter;
+
+service Greeter {
+  option (dev.restate.ext.service_type) = SINGLETON;
+
+  rpc Greet(GreetingRequest) returns (GreetingResponse);
+}
+
+message GreetingRequest {
+  string person = 1;
+}
+
+message GreetingResponse {
+  string greeting = 1;
+}
