+++ crates/schema-impl/tests/pb/generic/greeter.proto
