+++ crates/service-protocol/src/discovery.rs
@@ -20,17 +20,10 @@ use hyper::http::uri::PathAndQuery;
 use hyper::http::{HeaderName, HeaderValue};
 use hyper::{Body, HeaderMap};
 use prost::{DecodeError, Message};
-use prost_reflect::{
-    Cardinality, DescriptorError, DescriptorPool, ExtensionDescriptor, FieldDescriptor, Kind,
-    MethodDescriptor, ServiceDescriptor,
-};
+use prost_reflect::{DescriptorError, DescriptorPool};
 
 use codederror::CodedError;
-use restate_errors::{warn_it, META0001, META0002, META0003, META0007, META0008};
-use restate_schema_api::discovery::{
-    DiscoveredInstanceType, DiscoveredMethodMetadata, FieldAnnotation, KeyStructure,
-    ServiceRegistrationRequest,
-};
+use restate_errors::{warn_it, META0003};
 use restate_schema_api::endpoint::ProtocolType;
 use restate_service_client::{
     Parts, Request, ServiceClient, ServiceClientError, ServiceEndpointAddress,
@@ -42,19 +35,6 @@ use restate_types::retries::{RetryIter, RetryPolicy};
 // https://github.com/rust-lang/rust/issues/40543#issuecomment-1212981256
 #[allow(clippy::declare_interior_mutable_const)]
 const APPLICATION_PROTO: HeaderValue = HeaderValue::from_static("application/proto");
-
-const RESTATE_SERVICE_NAME_PREFIX: &str = "dev.restate.";
-const GRPC_SERVICE_NAME_PREFIX: &str = "grpc.";
-const SERVICE_TYPE_EXT: &str = "dev.restate.ext.service_type";
-const FIELD_EXT: &str = "dev.restate.ext.field";
-
-const UNKEYED_SERVICE_EXT: i32 = 0;
-const KEYED_SERVICE_EXT: i32 = 1;
-const SINGLETON_SERVICE_EXT: i32 = 2;
-
-const KEY_FIELD_EXT: i32 = 0;
-const EVENT_PAYLOAD_FIELD_EXT: i32 = 1;
-const EVENT_METADATA_FIELD_EXT: i32 = 2;
 
 const DISCOVER_PATH: &str = "/discover";
 
@@ -147,73 +127,14 @@ impl DiscoverEndpoint {
 
 #[derive(Debug)]
 pub struct DiscoveredEndpointMetadata {
-    pub services: Vec<ServiceRegistrationRequest>,
+    pub services: Vec<String>,
     pub descriptor_pool: DescriptorPool,
     pub protocol_type: ProtocolType,
 }
 
 #[derive(Debug, thiserror::Error, CodedError)]
 pub enum ServiceDiscoveryError {
-    // User errors
-    #[error("bad uri '{0}'. The uri must contain either `http` or `https` scheme, a valid authority and can contain a path where the service endpoint is exposed.")]
-    #[code(unknown)]
-    BadUri(String),
-    #[error("cannot find the dev.restate.ext.service_type extension in the descriptor of service '{0}'. You must annotate a service using the dev.restate.ext.service_type extension to specify whether your service is KEYED, UNKEYED or SINGLETON")]
-    #[code(META0001)]
-    MissingServiceTypeExtension(String),
-    #[error("the service '{0}' is keyed but has no methods. You must specify at least one method")]
-    #[code(META0001)]
-    KeyedServiceWithoutMethods(String),
-    #[error("the service name '{0}' is reserved. Service name should must not start with 'dev.restate' or 'grpc'")]
-    #[code(META0001)]
-    ServiceNameReserved(String),
-    #[error(
-        "error when trying to parse the key of service method '{}' with input type '{}'. No key field found",
-        MethodDescriptor::full_name(.0),
-        MethodDescriptor::input(.0).full_name()
-    )]
-    #[code(META0002)]
-    MissingKeyField(MethodDescriptor),
-    #[error(
-        "error when trying to parse the annotation {1:?} of service method '{}' with input type '{}'. More than one field annotated with the same annotation found",
-        MethodDescriptor::full_name(.0),
-        MethodDescriptor::input(.0).full_name()
-    )]
-    #[code(META0007)]
-    MoreThanOneAnnotatedField(MethodDescriptor, FieldAnnotation),
-    #[error(
-        "error when trying to parse the key of service method '{}' with input type '{}'. Bad key field type",
-        MethodDescriptor::full_name(.0),
-        MethodDescriptor::input(.0).full_name()
-    )]
-    #[code(META0002)]
-    BadKeyFieldType(MethodDescriptor),
-    #[error(
-        "error when trying to parse the key of service method '{}' with input type '{}'. The key type is different from other methods key types",
-        MethodDescriptor::full_name(.0),
-        MethodDescriptor::input(.0).full_name()
-    )]
-    #[code(META0002)]
-    DifferentKeyTypes(MethodDescriptor),
-    #[error(
-        "error when parsing the annotation EVENT_PAYLOAD of service method '{}' with input type '{}'. Bad type",
-        MethodDescriptor::full_name(.0),
-        MethodDescriptor::input(.0).full_name()
-    )]
-    #[code(META0008)]
-    BadEventPayloadFieldType(MethodDescriptor),
-    #[error(
-    "error when parsing the annotation EVENT_METADATA of service method '{}' with input type '{}'. Bad type",
-        MethodDescriptor::full_name(.0),
-        MethodDescriptor::input(.0).full_name()
-    )]
-    #[code(META0008)]
-    BadEventMetadataFieldType(MethodDescriptor),
-
     // Errors most likely related to SDK bugs
-    #[error("cannot find service '{0}' in descriptor set. This might be a symptom of an SDK bug, or of the build tool/pipeline used to generate the descriptor")]
-    #[code(unknown)]
-    ServiceNotFoundInDescriptor(String),
     #[error("received a bad response from the SDK: {0}. This might be a symptom of an SDK bug")]
     #[code(unknown)]
     BadResponse(&'static str),
@@ -223,9 +144,6 @@ pub enum ServiceDiscoveryError {
     #[error("received a bad response from the SDK with a descriptor set that cannot be reconstructed: {0}. This might be a symptom of an SDK bug")]
     #[code(unknown)]
     Descriptor(#[from] DescriptorError),
-    #[error("bad or missing Restate dependency in the descriptor pool. This might be a symptom of an SDK bug")]
-    #[code(unknown)]
-    BadOrMissingRestateDependencyInDescriptor,
 
     // Network related retryable errors
     #[error("retry limit exhausted. Last bad status code: {0}")]
@@ -234,23 +152,6 @@ pub enum ServiceDiscoveryError {
     #[error("retry limit exhausted. Last client error: {0}")]
     #[code(META0003)]
     Client(#[from] ServiceClientError),
-}
-
-impl ServiceDiscoveryError {
-    pub fn is_user_error(&self) -> bool {
-        matches!(
-            self,
-            ServiceDiscoveryError::BadUri(_)
-                | ServiceDiscoveryError::MissingServiceTypeExtension(_)
-                | ServiceDiscoveryError::KeyedServiceWithoutMethods(_)
-                | ServiceDiscoveryError::MissingKeyField(_)
-                | ServiceDiscoveryError::MoreThanOneAnnotatedField(_, _)
-                | ServiceDiscoveryError::BadKeyFieldType(_)
-                | ServiceDiscoveryError::DifferentKeyTypes(_)
-                | ServiceDiscoveryError::BadEventPayloadFieldType(_)
-                | ServiceDiscoveryError::BadEventMetadataFieldType(_)
-        )
-    }
 }
 
 impl ServiceDiscovery {
@@ -285,57 +186,8 @@ impl ServiceDiscovery {
         let response: pb::ServiceDiscoveryResponse = pb::ServiceDiscoveryResponse::decode(body)?;
         let descriptor_pool = DescriptorPool::decode(patch_built_in_descriptors(response.files)?)?;
 
-        // Find the Restate extensions in the DescriptorPool.
-        // If they're not available, the descriptor pool is incomplete/doesn't contain the restate dependencies.
-        let restate_service_type_extension = descriptor_pool
-            .get_extension_by_name(SERVICE_TYPE_EXT)
-            .ok_or(ServiceDiscoveryError::BadOrMissingRestateDependencyInDescriptor)?;
-        let restate_key_extension = descriptor_pool
-            .get_extension_by_name(FIELD_EXT)
-            .ok_or(ServiceDiscoveryError::BadOrMissingRestateDependencyInDescriptor)?;
-
-        // Collect all the service descriptors
-        let service_descriptors = response
-            .services
-            .into_iter()
-            .map(|svc| {
-                descriptor_pool
-                    .get_service_by_name(&svc)
-                    .ok_or(ServiceDiscoveryError::ServiceNotFoundInDescriptor(svc))
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-
-        // Infer service instance type
-        let mut services = Vec::with_capacity(service_descriptors.len());
-        for svc_desc in service_descriptors {
-            let svc_name = svc_desc.full_name();
-            if svc_name.starts_with(RESTATE_SERVICE_NAME_PREFIX)
-                || svc_name.starts_with(GRPC_SERVICE_NAME_PREFIX)
-            {
-                return Err(ServiceDiscoveryError::ServiceNameReserved(
-                    svc_desc.full_name().to_string(),
-                ));
-            }
-            let mut methods = HashMap::with_capacity(svc_desc.methods().len());
-
-            let instance_type = infer_service_type(
-                &svc_desc,
-                &restate_service_type_extension,
-                &restate_key_extension,
-                &mut methods,
-            )?;
-
-            infer_event_fields_annotations(&svc_desc, &restate_key_extension, &mut methods)?;
-
-            services.push(ServiceRegistrationRequest {
-                name: svc_name.to_string(),
-                instance_type,
-                methods,
-            })
-        }
-
         Ok(DiscoveredEndpointMetadata {
-            services,
+            services: response.services,
             descriptor_pool,
             protocol_type: match pb::ProtocolMode::try_from(response.protocol_mode) {
                 Ok(pb::ProtocolMode::BidiStream) => ProtocolType::BidiStream,
@@ -398,229 +250,6 @@ impl ServiceDiscovery {
             }
         }
     }
-}
-
-pub fn infer_service_type(
-    desc: &ServiceDescriptor,
-    restate_service_type_ext: &ExtensionDescriptor,
-    restate_field_ext: &ExtensionDescriptor,
-    methods: &mut HashMap<String, DiscoveredMethodMetadata>,
-) -> Result<DiscoveredInstanceType, ServiceDiscoveryError> {
-    if !desc.options().has_extension(restate_service_type_ext) {
-        return Err(ServiceDiscoveryError::MissingServiceTypeExtension(
-            desc.full_name().to_string(),
-        ));
-    }
-
-    let service_instance_type = desc
-        .options()
-        .get_extension(restate_service_type_ext)
-        .as_enum_number()
-        // This can happen only if the restate dependency is bad?
-        .ok_or_else(|| ServiceDiscoveryError::BadOrMissingRestateDependencyInDescriptor)?;
-
-    match service_instance_type {
-        UNKEYED_SERVICE_EXT => Ok(DiscoveredInstanceType::Unkeyed),
-        KEYED_SERVICE_EXT => infer_keyed_service_type(desc, restate_field_ext, methods),
-        SINGLETON_SERVICE_EXT => Ok(DiscoveredInstanceType::Singleton),
-        _ => Err(ServiceDiscoveryError::BadOrMissingRestateDependencyInDescriptor),
-    }
-}
-
-pub fn infer_keyed_service_type(
-    desc: &ServiceDescriptor,
-    restate_field_ext: &ExtensionDescriptor,
-    methods: &mut HashMap<String, DiscoveredMethodMetadata>,
-) -> Result<DiscoveredInstanceType, ServiceDiscoveryError> {
-    if desc.methods().len() == 0 {
-        return Err(ServiceDiscoveryError::KeyedServiceWithoutMethods(
-            desc.full_name().to_string(),
-        ));
-    }
-
-    // Parse the key from the first method
-    let first_method = desc.methods().next().unwrap();
-    let first_key_field_descriptor = resolve_key_field(&first_method, restate_field_ext)?;
-
-    // Generate the KeyStructure out of it
-    let key_structure = infer_key_structure(&first_key_field_descriptor);
-    methods
-        .entry(first_method.name().to_string())
-        .or_default()
-        .input_fields_annotations
-        .insert(FieldAnnotation::Key, first_key_field_descriptor.number());
-
-    // Now parse the next methods
-    for method_desc in desc.methods().skip(1) {
-        let key_field_descriptor = resolve_key_field(&method_desc, restate_field_ext)?;
-
-        // Validate every method has the same key field type
-        if key_field_descriptor.kind() != first_key_field_descriptor.kind() {
-            return Err(ServiceDiscoveryError::DifferentKeyTypes(method_desc));
-        }
-
-        methods
-            .entry(method_desc.name().to_string())
-            .or_default()
-            .input_fields_annotations
-            .insert(FieldAnnotation::Key, key_field_descriptor.number());
-    }
-
-    Ok(DiscoveredInstanceType::Keyed(key_structure))
-}
-
-fn infer_key_structure(field_descriptor: &FieldDescriptor) -> KeyStructure {
-    match field_descriptor.kind() {
-        Kind::Message(message_descriptor) => KeyStructure::Nested(
-            message_descriptor
-                .fields()
-                .map(|f| (f.number(), infer_key_structure(&f)))
-                .collect(),
-        ),
-        _ => KeyStructure::Scalar,
-    }
-}
-
-fn resolve_key_field(
-    method_descriptor: &MethodDescriptor,
-    restate_field_extension: &ExtensionDescriptor,
-) -> Result<FieldDescriptor, ServiceDiscoveryError> {
-    let field_descriptor = match get_annotated_field(
-        method_descriptor,
-        restate_field_extension,
-        KEY_FIELD_EXT,
-        FieldAnnotation::Key,
-    )? {
-        Some(f) => f,
-        None => {
-            return Err(ServiceDiscoveryError::MissingKeyField(
-                method_descriptor.clone(),
-            ));
-        }
-    };
-
-    if field_descriptor.kind() != Kind::String {
-        return Err(ServiceDiscoveryError::BadKeyFieldType(
-            method_descriptor.clone(),
-        ));
-    }
-
-    Ok(field_descriptor)
-}
-
-fn infer_event_fields_annotations(
-    service_desc: &ServiceDescriptor,
-    restate_field_ext: &ExtensionDescriptor,
-    methods: &mut HashMap<String, DiscoveredMethodMetadata>,
-) -> Result<(), ServiceDiscoveryError> {
-    for method_desc in service_desc.methods() {
-        let method_meta = methods.entry(method_desc.name().to_string()).or_default();
-
-        // Infer event annotations
-        if let Some(field_desc) = resolve_event_payload_field(&method_desc, restate_field_ext)? {
-            method_meta
-                .input_fields_annotations
-                .insert(FieldAnnotation::EventPayload, field_desc.number());
-        }
-        if let Some(field_desc) = resolve_event_metadata_field(&method_desc, restate_field_ext)? {
-            method_meta
-                .input_fields_annotations
-                .insert(FieldAnnotation::EventMetadata, field_desc.number());
-        }
-    }
-    Ok(())
-}
-
-fn resolve_event_payload_field(
-    method_descriptor: &MethodDescriptor,
-    restate_field_extension: &ExtensionDescriptor,
-) -> Result<Option<FieldDescriptor>, ServiceDiscoveryError> {
-    let field_descriptor = match get_annotated_field(
-        method_descriptor,
-        restate_field_extension,
-        EVENT_PAYLOAD_FIELD_EXT,
-        FieldAnnotation::EventPayload,
-    )? {
-        Some(f) => f,
-        None => return Ok(None),
-    };
-
-    // Validate type
-    if field_descriptor.kind() != Kind::String && field_descriptor.kind() != Kind::Bytes {
-        return Err(ServiceDiscoveryError::BadEventPayloadFieldType(
-            method_descriptor.clone(),
-        ));
-    }
-
-    Ok(Some(field_descriptor))
-}
-
-fn resolve_event_metadata_field(
-    method_descriptor: &MethodDescriptor,
-    restate_field_extension: &ExtensionDescriptor,
-) -> Result<Option<FieldDescriptor>, ServiceDiscoveryError> {
-    let field_descriptor = match get_annotated_field(
-        method_descriptor,
-        restate_field_extension,
-        EVENT_METADATA_FIELD_EXT,
-        FieldAnnotation::EventMetadata,
-    )? {
-        Some(f) => f,
-        None => return Ok(None),
-    };
-
-    // Validate type
-    if !is_map_with(&field_descriptor, Kind::String, Kind::String) {
-        return Err(ServiceDiscoveryError::BadEventMetadataFieldType(
-            method_descriptor.clone(),
-        ));
-    }
-
-    Ok(Some(field_descriptor))
-}
-
-fn get_annotated_field(
-    method_descriptor: &MethodDescriptor,
-    restate_field_extension: &ExtensionDescriptor,
-    extension_value: i32,
-    field_annotation: FieldAnnotation,
-) -> Result<Option<FieldDescriptor>, ServiceDiscoveryError> {
-    let message_desc = method_descriptor.input();
-    let mut iter = message_desc.fields().filter(|f| {
-        f.options().has_extension(restate_field_extension)
-            && f.options()
-                .get_extension(restate_field_extension)
-                .as_enum_number()
-                == Some(extension_value)
-    });
-
-    let field = iter.next();
-    if field.is_none() {
-        return Ok(None);
-    }
-
-    // Check there is only one
-    if iter.next().is_some() {
-        return Err(ServiceDiscoveryError::MoreThanOneAnnotatedField(
-            method_descriptor.clone(),
-            field_annotation,
-        ));
-    }
-
-    Ok(field)
-}
-
-// Expanded version of FieldDescriptor::is_map
-fn is_map_with(field_descriptor: &FieldDescriptor, key_kind: Kind, value_kind: Kind) -> bool {
-    field_descriptor.cardinality() == Cardinality::Repeated
-        && match field_descriptor.kind() {
-            Kind::Message(message) => {
-                message.is_map_entry()
-                    && message.map_entry_key_field().kind() == key_kind
-                    && message.map_entry_value_field().kind() == value_kind
-            }
-            _ => false,
-        }
 }
 
 // This function patches the built-in descriptors, to fix https://github.com/restatedev/restate/issues/687
